Move the platform agent surface into druks/mcp/gateway
Adds druks/mcp/gateway as the inbound platform surface beside the outbound
MCP-server registry: services (gate read/answer, run cancel, bounded agent-call
detail, and usage), the {code, message, retryable} AgentApiError base and its
codes, the wire schemas, and the tags=["agent"] routes for /api/agent/* and
/api/usage/agent. The durable and usage agent halves collapse into
gateway/services; their platform schemas and the exception base move with them,
api/app.py mounts the gateway router and renders AgentApiError off it, and
build's errors now subclass the moved base.

TextSlice, AgentCallSummary and UsageHistoryPoint stay in their domains —
read_slice, RunSummary and UsageHarnessHistory still depend on them and nothing
in a domain may import the gateway — so the gateway imports them down.
