+++ backend/druks/durable/exceptions.py
@@ -1,7 +1,5 @@
 from typing import ClassVar
 
-from druks.exceptions import AgentApiError
-
 
 class FatalError(Exception):
     """End the run as failed on purpose: the message becomes the run's recorded
@@ -33,58 +31,3 @@ class SubjectlessGate(FatalError):
             "to notify someone directly"
         )
         self.gate = gate
-
-
-class RunNotFound(AgentApiError):
-    status_code = 404
-    code = "RUN_NOT_FOUND"
-
-    def __init__(self, run_id: str) -> None:
-        super().__init__(f"No run {run_id}.")
-
-
-class GateNotOpen(AgentApiError):
-    status_code = 409
-    code = "GATE_NOT_OPEN"
-
-    def __init__(self, run_id: str) -> None:
-        super().__init__(f"Run {run_id} is not parked on a gate.")
-
-
-class GateRoundStale(AgentApiError):
-    status_code = 409
-    code = "GATE_ROUND_STALE"
-    retryable = True
-
-    def __init__(self, run_id: str) -> None:
-        super().__init__(
-            f"Run {run_id} has re-parked since the parked_at you read; fetch the gate again."
-        )
-
-
-class GateNotAnswerable(AgentApiError):
-    status_code = 409
-    code = "GATE_NOT_ANSWERABLE"
-
-    def __init__(self, run_id: str) -> None:
-        super().__init__(f"The gate on run {run_id} is answered on its source, not through here.")
-
-
-class InvalidGateAnswer(AgentApiError):
-    code = "INVALID_GATE_ANSWER"
-
-
-class AgentCallNotFound(AgentApiError):
-    status_code = 404
-    code = "AGENT_CALL_NOT_FOUND"
-
-    def __init__(self, call_id: str) -> None:
-        super().__init__(f"No agent call {call_id}.")
-
-
-class RunNotActive(AgentApiError):
-    status_code = 409
-    code = "RUN_NOT_ACTIVE"
-
-    def __init__(self, run_id: str) -> None:
-        super().__init__(f"Run {run_id} already ended.")
