+++ backend/druks/durable/schemas.py
@@ -290,43 +290,3 @@ class RunSummary(BaseResponse):
             account_username=run.account.username,
             agent_calls=[AgentCallSummary.from_call(call) for call in calls],
         )
-
-
-class ArtifactChunk(BaseResponse):
-    # A call's renderable output, head-bounded; page the rest through the
-    # call's transcript files route.
-    call_id: str
-    kind: str
-    title: str
-    chunk: TextSlice
-
-
-class GateView(BaseResponse):
-    # Everything needed to answer a parked run in one read: the ask, the
-    # artifact under review, the reply's JSON Schema, and parked_at — the park
-    # identity answer_gate must echo back.
-    run_id: str
-    gate: str
-    parked_at: datetime
-    ask: dict[str, Any]
-    artifact: ArtifactChunk | None = None
-    reply_schema: dict[str, Any]
-
-
-class GateAnswerResult(BaseResponse):
-    run_id: str
-    parked_at: datetime
-    result: Literal["answered", "already_answered"]
-
-
-class AgentCallDetail(BaseResponse):
-    run_id: str
-    call: AgentCallSummary
-    transcript: TextSlice
-    stderr: TextSlice
-    artifact: ArtifactChunk | None = None
-
-
-class CancelRunResult(BaseResponse):
-    run_id: str
-    result: Literal["cancelled", "already_cancelled"]
