+++ backend/druks/mcp/gateway/routes.py
@@ -1,13 +1,21 @@
 from typing import Annotated
 
-from fastapi import APIRouter, Body
+from fastapi import APIRouter, Body, Depends
 from pydantic import AwareDatetime, BaseModel, ConfigDict, Field
 from pydantic.alias_generators import to_camel
 
-from druks.durable import agent
-from druks.durable.schemas import AgentCallDetail, CancelRunResult, GateAnswerResult, GateView
+from druks.accounts.dependencies import current_account
+from druks.accounts.models import Account
+from druks.mcp.gateway import services
+from druks.mcp.gateway.schemas import (
+    AgentCallDetail,
+    AgentUsage,
+    CancelRunResult,
+    GateAnswerResult,
+    GateView,
+)
 
-router = APIRouter(prefix="/api/agent", tags=["agent"])
+router = APIRouter(tags=["agent"])
 
 
 class AnswerGateRequest(BaseModel):
@@ -23,23 +31,23 @@ class AnswerGateRequest(BaseModel):
 
 
 @router.get(
-    "/gates/{run_id}",
+    "/api/agent/gates/{run_id}",
     operation_id="get_gate",
     response_model=GateView,
     response_model_by_alias=True,
 )
 async def get_gate(run_id: str) -> GateView:
-    return agent.get_gate(run_id)
+    return services.get_gate(run_id)
 
 
 @router.post(
-    "/gates/{run_id}/answer",
+    "/api/agent/gates/{run_id}/answer",
     operation_id="answer_gate",
     response_model=GateAnswerResult,
     response_model_by_alias=True,
 )
 async def answer_gate(run_id: str, body: AnswerGateRequest) -> GateAnswerResult:
-    return await agent.answer_gate(
+    return await services.answer_gate(
         run_id,
         parked_at=body.parked_at,
         control=body.control,
@@ -49,17 +57,17 @@ async def answer_gate(run_id: str, body: AnswerGateRequest) -> GateAnswerResult:
 
 
 @router.get(
-    "/agent-calls/{call_id}",
+    "/api/agent/agent-calls/{call_id}",
     operation_id="get_agent_call",
     response_model=AgentCallDetail,
     response_model_by_alias=True,
 )
 async def get_agent_call(call_id: str) -> AgentCallDetail:
-    return agent.get_agent_call(call_id)
+    return services.get_agent_call(call_id)
 
 
 @router.post(
-    "/runs/{run_id}/cancel",
+    "/api/agent/runs/{run_id}/cancel",
     operation_id="cancel_run",
     response_model=CancelRunResult,
     response_model_by_alias=True,
@@ -67,4 +75,14 @@ async def get_agent_call(call_id: str) -> AgentCallDetail:
 async def cancel_run(
     run_id: str, reason: Annotated[str, Body(embed=True, min_length=1, max_length=500)]
 ) -> CancelRunResult:
-    return await agent.cancel_run(run_id, reason=reason)
+    return await services.cancel_run(run_id, reason=reason)
+
+
+@router.get(
+    "/api/usage/agent",
+    operation_id="get_usage",
+    response_model=AgentUsage,
+    response_model_by_alias=True,
+)
+async def get_usage(account: Account = Depends(current_account)) -> AgentUsage:
+    return services.get_usage(account)
